Port to hal-0.7
