+++ Cargo.toml
@@ -2,12 +2,14 @@
 name = "gfx-ocean"
 version = "0.1.0"
 edition = "2018"
+publish = false
 
 [features]
 default = []
 dx12 = ["gfx-backend-dx12"]
 vulkan = ["gfx-backend-vulkan"]
 metal = ["gfx-backend-metal"]
+gl = ["gfx-backend-gl"]
 
 [lib]
 name = "ocean"
@@ -23,17 +25,28 @@ required-features = ["metal"]
 name = "ocean"
 
 [dependencies]
-env_logger = "0.7"
+env_logger = "0.8"
 log = "0.4"
-winit = "0.22"
-nalgebra-glm = "0.4"
+winit = "0.24"
+glm = { package = "nalgebra-glm", version = "0.4" }
 bincode = "1.1"
 
-gfx-hal = "0.6"
-gfx-auxil = "0.5"
-gfx-backend-vulkan = { version = "0.6", features = ["x11", "xcb"], optional = true }
+hal = { package = "gfx-hal", version = "0.7" }
+gfx-auxil = "0.8"
+gfx-backend-empty = "0.7"
+gfx-backend-vulkan = { version = "0.7", optional = true }
 [target.'cfg(windows)'.dependencies]
-gfx-backend-dx12 = { version = "0.6", optional = true }
+gfx-backend-dx12 = { version = "0.7", optional = true }
 [target.'cfg(any(target_os = "macos", all(target_os = "ios", target_arch = "aarch64")))'.dependencies]
-gfx-backend-metal = { version = "0.6", optional = true }
+gfx-backend-metal = { version = "0.7", optional = true }
 objc = "0.2.7"
+[target.'cfg(all(unix, not(target_os = "macos"), not(target_os = "ios")))'.dependencies]
+gfx-backend-gl = { version = "0.7", optional = true }
+
+[patch.crates-io]
+#gfx-hal = { path = "../gfx/src/hal" }
+#gfx-backend-empty = { path = "../gfx/src/backend/empty" }
+#gfx-backend-vulkan = { path = "../gfx/src/backend/vulkan" }
+#gfx-backend-gl = { path = "../gfx/src/backend/gl" }
+#gfx-backend-dx12 = { path = "../gfx/src/backend/dx12" }
+#gfx-backend-metal = { path = "../gfx/src/backend/metal" }
