+++ src/fft.rs
@@ -1,26 +1,22 @@
-use crate::back::Backend as B;
-use gfx_hal::{
+use hal::{
     device::Device as _,
     pso::{self, DescriptorPool as _},
-    Backend,
 };
-use std::io::Cursor;
+use std::{io::Cursor, iter};
 
-pub struct Fft {
-    pub cs_fft_row: <B as Backend>::ShaderModule,
-    pub cs_fft_col: <B as Backend>::ShaderModule,
-    pub set_layout: <B as Backend>::DescriptorSetLayout,
-    pub layout: <B as Backend>::PipelineLayout,
-    pub row_pass: <B as Backend>::ComputePipeline,
-    pub col_pass: <B as Backend>::ComputePipeline,
-    pub desc_sets: Vec<<B as Backend>::DescriptorSet>,
-    pub pool: <B as Backend>::DescriptorPool,
+pub struct Fft<B: hal::Backend> {
+    pub cs_fft_row: B::ShaderModule,
+    pub cs_fft_col: B::ShaderModule,
+    pub set_layout: B::DescriptorSetLayout,
+    pub layout: B::PipelineLayout,
+    pub row_pass: B::ComputePipeline,
+    pub col_pass: B::ComputePipeline,
+    pub desc_sets: Vec<B::DescriptorSet>,
+    pub pool: B::DescriptorPool,
 }
 
-impl Fft {
-    pub unsafe fn init(
-        device: &<B as Backend>::Device,
-    ) -> Result<Self, Box<dyn std::error::Error>> {
+impl<B: hal::Backend> Fft<B> {
+    pub unsafe fn init(device: &B::Device) -> Result<Self, Box<dyn std::error::Error>> {
         let cs_fft_row = device.create_shader_module(&gfx_auxil::read_spirv(Cursor::new(
             &include_bytes!("../shader/spv/fft_row.comp.spv")[..],
         ))?)?;
@@ -29,7 +25,7 @@ impl Fft {
         ))?)?;
 
         let set_layout = device.create_descriptor_set_layout(
-            &[pso::DescriptorSetLayoutBinding {
+            iter::once(pso::DescriptorSetLayoutBinding {
                 binding: 0,
                 ty: pso::DescriptorType::Buffer {
                     ty: pso::BufferDescriptorType::Storage { read_only: false },
@@ -40,13 +36,13 @@ impl Fft {
                 count: 1,
                 stage_flags: pso::ShaderStageFlags::COMPUTE,
                 immutable_samplers: false,
-            }],
-            &[],
+            }),
+            iter::empty(),
         )?;
 
         let mut pool = device.create_descriptor_pool(
             3,
-            &[pso::DescriptorRangeDesc {
+            iter::once(pso::DescriptorRangeDesc {
                 ty: pso::DescriptorType::Buffer {
                     ty: pso::BufferDescriptorType::Storage { read_only: false },
                     format: pso::BufferDescriptorFormat::Structured {
@@ -54,40 +50,42 @@ impl Fft {
                     },
                 },
                 count: 3,
-            }],
+            }),
             pso::DescriptorPoolCreateFlags::empty(),
         )?;
 
-        let mut desc_sets = Vec::<<B as Backend>::DescriptorSet>::with_capacity(3);
-        pool.allocate(vec![&set_layout, &set_layout, &set_layout], &mut desc_sets)?;
-        let layout = device.create_pipeline_layout(Some(&set_layout), &[])?;
-        let (row_pass, col_pass) = {
-            let mut pipelines = device.create_compute_pipelines(
-                &[
-                    pso::ComputePipelineDesc::new(
-                        pso::EntryPoint {
-                            entry: "main",
-                            module: &cs_fft_row,
-                            specialization: pso::Specialization::default(),
-                        },
-                        &layout,
-                    ),
-                    pso::ComputePipelineDesc::new(
-                        pso::EntryPoint {
-                            entry: "main",
-                            module: &cs_fft_col,
-                            specialization: pso::Specialization::default(),
-                        },
-                        &layout,
-                    ),
-                ],
+        let mut desc_sets = Vec::<B::DescriptorSet>::with_capacity(3);
+        pool.allocate(
+            vec![&set_layout, &set_layout, &set_layout].into_iter(),
+            &mut desc_sets,
+        )?;
+        let layout = device.create_pipeline_layout(iter::once(&set_layout), iter::empty())?;
+        let row_pass = device
+            .create_compute_pipeline(
+                &pso::ComputePipelineDesc::new(
+                    pso::EntryPoint {
+                        entry: "main",
+                        module: &cs_fft_row,
+                        specialization: pso::Specialization::default(),
+                    },
+                    &layout,
+                ),
                 None,
-            );
-
-            let row_pass = pipelines.remove(0).unwrap();
-            let col_pass = pipelines.remove(0).unwrap();
-            (row_pass, col_pass)
-        };
+            )
+            .unwrap();
+        let col_pass = device
+            .create_compute_pipeline(
+                &pso::ComputePipelineDesc::new(
+                    pso::EntryPoint {
+                        entry: "main",
+                        module: &cs_fft_col,
+                        specialization: pso::Specialization::default(),
+                    },
+                    &layout,
+                ),
+                None,
+            )
+            .unwrap();
 
         Ok(Fft {
             cs_fft_row,
@@ -101,7 +99,7 @@ impl Fft {
         })
     }
 
-    pub unsafe fn destroy(self, device: &<B as Backend>::Device) {
+    pub unsafe fn destroy(self, device: &B::Device) {
         device.destroy_shader_module(self.cs_fft_row);
         device.destroy_shader_module(self.cs_fft_col);
         device.destroy_descriptor_set_layout(self.set_layout);
